backend/requirements.txt
fastapi==0.115.12
importlib-metadata==8.0.0
inflect==7.3.1
jaraco.collections==5.1.0
llama-cpp-python==0.3.9
packaging==24.2
pip-chill==1.0.3
platformdirs==4.2.2
psycopg2-binary==2.9.10
sqlalchemy==2.0.40
tomli==2.0.1
uvicorn
